requirements.txt
wheel
pycrypto
argparse
configparser
Wand